Thread manager is realized; added ErrorCollector

+++ gvEngine/ThreadManager.h
@@ -15,8 +15,8 @@ namespace gv
 			std::shared_ptr<IRenderer> _renderer;
 			std::shared_ptr<IInputController> _inputController;		
 
-			std::chrono::duration<double, std::milli> _inputCheckInterval;
-			std::chrono::duration<double, std::milli> _renderLoopInterval;
+			std::chrono::duration<int, std::milli> _inputCheckInterval;
+			std::chrono::duration<int, std::milli> _renderLoopInterval;
 			std::chrono::system_clock::time_point _lastRenderTime;
 
 
@@ -24,8 +24,8 @@ namespace gv
 			///all intervals in ms
 			ThreadManager(std::shared_ptr<IRenderer> renderer, 
 				std::shared_ptr<IInputController> inputController, 
-				float inputCheckInterval,
-				float renderLoopInterval);
+				unsigned int inputCheckInterval,
+				unsigned int renderLoopInterval);
 			void start();
 		};
 	}
